+++ apps/dine-pleiepenger/src/components/lenker/SkrivTilOssLenke.tsx
@@ -1,6 +1,6 @@
 import { Link } from '@navikt/ds-react';
-import { ExternalLink } from '@navikt/ds-icons';
 import { browserEnv } from '../../utils/env';
+import { ExternalLinkIcon } from '@navikt/aksel-icons';
 
 interface Props {
     tekst: string;
@@ -8,7 +8,7 @@ interface Props {
 
 const SkrivTilOssLenke = ({ tekst }: Props) => (
     <Link href={browserEnv.NEXT_PUBLIC_SKRIV_TIL_OSS_URL}>
-        {tekst} <ExternalLink className="ml-1" role="presentation" />
+        {tekst} <ExternalLinkIcon className="ml-1" role="presentation" />
     </Link>
 );