+++ apps/omsorgsdager-kalkulator/components/sif-formik/typed-formik-form/TypedFormikForm.tsx
@@ -1,11 +1,11 @@
 import { Button } from '@navikt/ds-react';
 import React, { createContext, useEffect, useRef, useState } from 'react';
-import { Back, Next } from '@navikt/ds-icons';
 import { FieldInputProps, FormikProps, useFormikContext } from 'formik';
 import FormikValidationErrorSummary from '../formik-validation-error-summary/FormikValidationErrorSummary';
 import ButtonRow from '../helpers/button-row/ButtonRow';
 import { CancelButtonTypes, CustomFormErrorHandler, ErrorTypeChecker, FieldErrorHandler, FormError } from '../types';
 import { getErrorForField, isValidationErrorsVisible } from '../utils/typedFormErrorUtils';
+import { ChevronLeftIcon, ChevronRightIcon } from '@navikt/aksel-icons';
 
 export interface TypedFormikFormProps<FormValues, ErrorType> {
     children: React.ReactNode;
@@ -182,7 +182,10 @@ function TypedFormikForm<FormValues, ErrorType>({
                                     data-testid="typedFormikForm-goBackButton"
                                     icon={
                                         showButtonArrows ? (
-                                            <Back aria-hidden className="typedFormikForm__buttonLabel__icon" />
+                                            <ChevronLeftIcon
+                                                aria-hidden
+                                                className="typedFormikForm__buttonLabel__icon"
+                                            />
                                         ) : undefined
                                     }>
                                     {backButtonLabel || 'Forrige'}
@@ -198,7 +201,10 @@ function TypedFormikForm<FormValues, ErrorType>({
                                     iconPosition="right"
                                     icon={
                                         showButtonArrows ? (
-                                            <Next aria-hidden className="typedFormikForm__buttonLabel__icon" />
+                                            <ChevronRightIcon
+                                                aria-hidden
+                                                className="typedFormikForm__buttonLabel__icon"
+                                            />
                                         ) : undefined
                                     }>
                                     {submitButtonLabel || 'Neste'}