+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-ettersending
 
+## 3.49.28
+
+### Patch Changes
+
+- Erstatte ds-icons med aksel-icons
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.5.24
+    - @navikt/sif-common-soknad-ds@26.1.8
+    - @navikt/sif-common-forms-ds@21.0.26
+    - @navikt/sif-common-core-ds@10.3.25
+    - @navikt/sif-common-ui@0.11.26
+
 ## 3.49.27
 
 ### Patch Changes