+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @navikt/sif-common-ui
 
+## 0.11.26
+
+### Patch Changes
+
+- Erstatte ds-icons med aksel-icons
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.5.24
+    - @navikt/sif-common-core-ds@10.3.25
+
 ## 0.11.25
 
 ### Patch Changes